Move BlogMarkdown to renderer folder
