+++ app/blog/preview/[...slug]/page.tsx
@@ -7,7 +7,7 @@ import Title from "../../../../components/Title";
 import {getPreviewPostData} from "../../../../lib/blog/loadPreview";
 import {fetchAllImageProps} from "../../../../lib/blog/imagePropsFetcher";
 
-import BlogMarkdown from "../../[...slug]/BlogMarkdown";
+import BlogMarkdown from "../../renderer/BlogMarkdown";
 
 import styles from '../../../../styles/blog/blog.module.scss';
