+++ components/blog/BlogImage.tsx
@@ -4,7 +4,7 @@ import React, {CSSProperties, useState} from "react";
 import styles from "../../styles/blog/BlogImage.module.scss";
 import Image from "next/legacy/image";
 import Modal from "react-modal";
-import {parseInlineMarkdown} from "../../app/blog/[...slug]/BlogMarkdown";
+import {parseInlineMarkdown} from "../../app/blog/renderer/BlogMarkdown";
 import {BlogImageData} from "../../lib/blog/imagePropsFetcher";
 import {FontAwesomeIcon} from "@fortawesome/react-fontawesome";
 import {faCamera} from "@fortawesome/free-solid-svg-icons";
