+++ components/blog/article-parts/Socials.tsx
@@ -4,7 +4,7 @@ import {Tweet} from "react-twitter-widgets";
 import YouTube from "react-youtube";
 import ReactPlayer from "react-player";
 import React from "react";
-import {parseInlineMarkdown} from "../../../app/blog/[...slug]/BlogMarkdown";
+import {parseInlineMarkdown} from "../../../app/blog/renderer/BlogMarkdown";
 import {ArticleParts} from "../ArticleParts";
 
 export const Twitter: ArticleParts = (content) => {
